update decorator to not create decorated functions per request
While the existing implementation is simple it: creates a decorated
function per request, calls it, and then will discard it. Since the
functionality is straightforward and also encapsulated fairly well in
Django, this PR modifies the decorator to copy the original Django
decorator, and calls the same helper functions where possible.

Because the functionality is no longer simply wrapping Django's internal
implementation, but is still mostly the same, there have been tests
added to compare the requests generated by Django vs requests generated
by Django Request Framework/this library.

Finally, in order to be less confusing to library writers the call
signature of `etag_func` and `last_modified_func` has been updated to
match Django Request Framework and will pass a DRF Request object
instead of Django's native Request object::

  # from the original signature
  def original_etag_func(wsgi_request, *args, **kwargs): pass

  # to the updated signature
  def updated_etag_func(drf_self, drf_request, *args, **kwargs): pass

In order to ease the transition between the two signatures, there has
been a flag "use_self" added to the decorators, and a warning will be
emitted if use_self has not been set to True.